+++ cinema-service/pkgs/parser/pkg_parser.go
@@ -9,13 +9,13 @@ import (
 	"strings"
 )
 
-func ParamsInt(r *http.Request, prefix string) (int32, error) {
+func ParamsInt(r *http.Request, prefix string) (int64, error) {
 	idStr := strings.TrimPrefix(r.URL.Path, prefix)
 	val, err := strconv.ParseInt(idStr, 10, 32)
 	if err != nil {
 		return 0, fmt.Errorf("ID tidak valid: %v", err)
 	}
-	return int32(val), nil
+	return val, nil
 }
 
 func Body[T any](r io.Reader, body T) (T, error) {
