+++ rest-client/internal/app/request/Access.go
@@ -8,17 +8,17 @@ type Permission struct {
 
 type Role struct {
 	Name         string  `validate:"required" json:"name"`
-	PermissionID []int32 `validate:"omitempty,dive,gt=0" json:"permission_id"`
+	PermissionID []int64 `validate:"omitempty,dive,gt=0" json:"permission_id"`
 }
 
 type GetRole struct {
-	ID         int32                       `json:"id"`
+	ID         int64                       `json:"id"`
 	Name       string                      `json:"name"`
 	Permission []repo.GetPermissionRoleRow `json:"permission"`
 }
 
 type GetClient struct {
-	ID    int32                   `json:"id"`
+	ID    int64                   `json:"id"`
 	Name  string                  `json:"name"`
 	Email string                  `json:"email"`
 	Role  []repo.GetRoleClientRow `json:"permission"`
