+++ rest-client/internal/app/service/genre_service.go
@@ -17,7 +17,7 @@ func ListGenre() ([]model.Genre, error) {
 	return data, nil
 }
 
-func GetGenre(id int32) (model.Genre, error) {
+func GetGenre(id int64) (model.Genre, error) {
 	url := fmt.Sprintf("/genre/%d", id)
 
 	data, err := gateway.GetCinema[model.Genre](url)
@@ -49,7 +49,7 @@ func UpdateGenre(body model.Genre) (model.Genre, error) {
 	return data, nil
 }
 
-func DeleteGenre(id int32) error {
+func DeleteGenre(id int64) error {
 	var url = fmt.Sprintf("/genre/delete/%d", id)
 
 	if err := gateway.DeleteCinema(url); err != nil {
